+++ src/ThemeContext.tsx
@@ -2,6 +2,7 @@ import React, { useState } from 'react'
 import { ThemeProvider as SCThemeProvider } from 'styled-components'
 import {  dark } from '@pancakeswap-libs/uikit'
 import { light } from 'constants/theme'
+
 const CACHE_KEY = 'IS_DARK'
 
 export interface ThemeContextType {
