+++ src/components/ConnectWalletButton/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { ButtonProps, useWalletModal} from '@pancakeswap-libs/uikit'
-import { Button } from '../Button';
 import useI18n from 'hooks/useI18n'
 import useAuth from 'hooks/useAuth'
+import { Button } from '../Button';
 
 const UnlockButton: React.FC<ButtonProps> = (props) => {
   const TranslateString = useI18n()
