+++ src/components/PageHeader/SlippageToleranceSetting.tsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
 import { Box, Flex, Input, Text } from '@pancakeswap-libs/uikit'
-import { Button } from '../../components/Button';
 import { useUserSlippageTolerance } from 'state/user/hooks'
 import QuestionHelper from '../QuestionHelper'
+import { Button } from '../Button';
 
 const MAX_SLIPPAGE = 5000
 const RISKY_SLIPPAGE_LOW = 50
