+++ src/components/PositionCard/index.tsx
@@ -1,7 +1,6 @@
 import React, { useState } from 'react'
 import { JSBI, Pair, Percent } from '@pancakeswap/sdk'
 import { Card as UIKitCard, CardBody, Text } from '@pancakeswap-libs/uikit'
-import { Button } from '../Button';
 import { darken } from 'polished'
 import { ChevronDown, ChevronUp } from 'react-feather'
 import { Link } from 'react-router-dom'
@@ -18,6 +17,7 @@ import CurrencyLogo from '../CurrencyLogo'
 import DoubleCurrencyLogo from '../DoubleLogo'
 import { RowBetween, RowFixed } from '../Row'
 import { Dots } from '../swap/styleds'
+import { Button } from '../Button';
 
 export const FixedHeightRow = styled(RowBetween)`
   height: 24px;
