+++ src/components/SearchModal/ListSelect.tsx
@@ -3,7 +3,6 @@ import { ArrowLeft } from 'react-feather'
 import { usePopper } from 'react-popper'
 import { useDispatch, useSelector } from 'react-redux'
 import { Text, ChevronDownIcon, CloseIcon } from '@pancakeswap-libs/uikit'
-import { Button } from '../Button';
 import styled from 'styled-components'
 import useI18n from 'hooks/useI18n'
 import { useFetchListCallback } from '../../hooks/useFetchListCallback'
@@ -21,6 +20,7 @@ import ListLogo from '../ListLogo'
 import QuestionHelper from '../QuestionHelper'
 import Row, { RowBetween } from '../Row'
 import { PaddedColumn, SearchInput, Separator, SeparatorDark } from './styleds'
+import { Button } from '../Button';
 
 const UnpaddedLinkStyledButton = styled(LinkStyledButton)`
   padding: 0;
