+++ src/components/TransactionConfirmationModal/TransactionErrorContent.tsx
@@ -1,10 +1,10 @@
 import React, { useContext } from 'react'
 import { ThemeContext } from 'styled-components'
 import { Text } from '@pancakeswap-libs/uikit'
-import { Button } from '../Button';
 import { AlertTriangle } from 'react-feather'
 import { AutoColumn } from '../Column'
 import { Wrapper, Section, BottomSection, ContentHeader } from './helpers'
+import { Button } from '../Button';
 
 type TransactionErrorContentProps = { message: string; onDismiss: () => void }
