+++ src/pages/AddLiquidity/ConfirmAddModalBottom.tsx
@@ -1,11 +1,11 @@
 import { Currency, CurrencyAmount, Fraction, Percent } from '@pancakeswap/sdk'
 import React from 'react'
 import { Text } from '@pancakeswap-libs/uikit'
-import { Button } from '../../components/Button';
 import { TranslateString } from 'utils/translateTextHelpers'
 import { RowBetween, RowFixed } from '../../components/Row'
 import CurrencyLogo from '../../components/CurrencyLogo'
 import { Field } from '../../state/mint/actions'
+import { Button } from '../../components/Button';
 
 export function ConfirmAddModalBottom({
   noLiquidity,
