+++ src/pages/RemoveLiquidity/index.tsx
@@ -5,13 +5,13 @@ import { Contract } from '@ethersproject/contracts'
 import { TransactionResponse } from '@ethersproject/providers'
 import { Currency, currencyEquals, ETHER, Percent } from '@pancakeswap/sdk'
 import { Flex, Text } from '@pancakeswap-libs/uikit'
-import { Button } from '../../components/Button';
 import { ArrowDown, Plus } from 'react-feather'
 import { RouteComponentProps } from 'react-router'
 
 import { BigNumber } from '@ethersproject/bignumber'
 import ConnectWalletButton from 'components/ConnectWalletButton'
 import useI18n from 'hooks/useI18n'
+import { Button } from '../../components/Button';
 import WETH from '../../constants/weth'
 import { AutoColumn, ColumnCenter } from '../../components/Column'
 import TransactionConfirmationModal, { ConfirmationModalContent } from '../../components/TransactionConfirmationModal'
